+++ packages/consumer/src/index.ts
@@ -1,14 +1,12 @@
-import type { VerifyCredentialResult } from "@idos-network/credentials/builder";
 import type {
   AvailableIssuerType,
-  IDDocumentType,
   idOSCredentialRecord as idOSCredential,
   VerifiableCredential,
-  VerifiableCredentialSubject,
 } from "@idos-network/credentials/types";
+import type { VerifyCredentialResult } from "@idos-network/credentials/verifier";
 import type { KwilSigner } from "@idos-network/kwil-js";
 
-import { type Credential, verifyCredential } from "@idos-network/credentials/builder";
+import { verifyCredential } from "@idos-network/credentials/verifier";
 import {
   createKgwAuthenticatedBlobGateway,
   createNodeKwilClient,
@@ -147,7 +145,7 @@ export class idOSConsumer {
     };
   }
 
-  async verifyCredential<K = VerifiableCredentialSubject>(
+  async verifyCredential<K>(
     credentials: VerifiableCredential<K>,
     issuers: AvailableIssuerType[],
   ): Promise<VerifyCredentialResult> {
@@ -158,10 +156,7 @@ export class idOSConsumer {
 export type {
   idOSCredential,
   idOSGrant,
-  Credential,
   VerifiableCredential,
-  VerifiableCredentialSubject,
   VerifyCredentialResult,
   AvailableIssuerType,
-  IDDocumentType,
 };