+++ packages/kwil-infra/tsdown.config.ts
@@ -2,6 +2,7 @@ import { defineConfig } from "tsdown";
 
 export default defineConfig({
   dts: true,
+  minify: true,
   entry: [
     "./src/index.ts",
     "./src/actions/index.ts",