+++ apps/idos-enclave/CHANGELOG.md
@@ -1,5 +1,24 @@
 # idos-enclave
 
+## 0.3.2
+
+### Patch Changes
+
+- Updated dependencies [1fb5dff]
+- Updated dependencies [1fb5dff]
+  - @idos-network/utils@2.0.1
+  - @idos-network/credentials@2.0.1
+  - @idos-network/enclave@2.0.1
+
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies [8d06f30]
+  - @idos-network/credentials@2.0.0
+  - @idos-network/enclave@2.0.0
+  - @idos-network/utils@2.0.0
+
 ## 0.3.0
 
 ### Minor Changes