+++ packages/kwil-infra/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @idos-network/kwil-infra
 
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies [1fb5dff]
+  - @idos-network/utils@2.0.1
+
+## 2.0.0
+
+### Patch Changes
+
+- @idos-network/utils@2.0.0
+
 ## 1.5.0
 
 ### Patch Changes